requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-ninja==1.3.0
pydantic==2.10.5
pydantic_core==2.27.2
sqlparse==0.5.3
typing_extensions==4.12.2
